+++ packages/jmap-rfc-types/lib/contracts/index.ts
@@ -1,0 +1,10 @@
+export type * from "./blob.ts";
+export type * from "./core.ts";
+export type * from "./email-submission.ts";
+export type * from "./email.ts";
+export type * from "./identity.ts";
+export type * from "./mailbox.ts";
+export type * from "./push-subscription.ts";
+export type * from "./search-snippet.ts";
+export type * from "./thread.ts";
+export type * from "./vacation-response.ts";